+++ src/main/java/com/example/rua/repository/RoleRepository.java
@@ -9,9 +9,12 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface RoleRepository extends JpaRepository<Roles,Long> {
+public interface RoleRepository extends JpaRepository<Roles,Integer> {
 
     //Select * from roles where name=?
     @Query("SELECT s FROM Roles s WHERE s.name=?1")
     Optional<Roles> findRoleByName(String name);
+
+
+
 }
